requirements.txt
#aiohttp==3.6.2
#async-timeout==3.0.1
#attrs==19.3.0
#autopep8==1.4.4
#cachetools==4.0.0
#certifi==2019.11.28
#chardet==3.0.4
#discord==1.0.1
discord.py==2.2.3
#elasticsearch==7.1.0
elasticsearch-dsl==7.4.1
#filetype==1.0.5
#google-cloud-compute==1.11
#google-cloud-vision==2.7.2
#idna==2.8
#multidict==4.7.2
#pillow==2.2.1
#pyasn1==0.4.8
#pyasn1-modules==0.2.7
#pycodestyle==2.5.0
#python-dateutil==2.8.1
#pytz==2019.3
#requests==2.22.0
#rsa==4.0
#six==1.13.0
#urllib3==1.25.7
#websockets==6.0
#yarl==1.4.2
